feat: my_local_modeというファイルの存在有無により，自分用のPC限定の設定を行うようにした

+++ linkDotfiles.sh
@@ -4,11 +4,15 @@
 # ln -sf ~/dotfiles/colors ~/.vim
 # ln -sf ~/dotfiles/ftdetect ~/.vim
 # ln -sf ~/dotfiles/indent ~/.vim
+
 ln -sf ~/dotfiles/.bash_profile ~/.bash_profile
 ln -sf ~/dotfiles/.bashrc ~/.bashrc
-ln -sf ~/dotfiles/.gemrc ~/.gemrc
 ln -sf ~/dotfiles/.gitconfig ~/.gitconfig
-ln -sf ~/dotfiles/.gvimrc ~/.gvimrc
-ln -sf ~/dotfiles/.irbrc ~/.irbrc
 ln -sf ~/dotfiles/.tmux.conf ~/.tmux.conf
 ln -sf ~/dotfiles/.vimrc ~/.vimrc
+
+if [ -s ~/dotfiles/my_local_mode* ]; then
+  ln -sf ~/dotfiles/.gemrc ~/.gemrc
+  ln -sf ~/dotfiles/.gvimrc ~/.gvimrc
+  ln -sf ~/dotfiles/.irbrc ~/.irbrc
+fi
